+++ src/main/java/com/api/ttoklip/domain/town/cart/repository/CartRepositoryCustom.java
@@ -34,4 +34,5 @@ public interface CartRepositoryCustom {
             final Long lastParty,
             final TownCriteria townCriteria
     );
+
 }
