+++ src/main/java/com/api/ttoklip/domain/town/cart/service/CartMemberService.java
@@ -7,6 +7,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+import java.util.Optional;
+
 
 @Service
 @RequiredArgsConstructor
@@ -19,6 +22,25 @@ public class CartMemberService {
     public void register(Cart cart) {
         CartMember newCartMember = CartMember.from(cart);
         cartMemberRepository.save(newCartMember);
+    }
 
+    public void delete(CartMember cartMember) {
+        cartMemberRepository.delete(cartMember);
+    }
+
+    public List<CartMember> findByCartId(Long cartId) {
+        return cartMemberRepository.findByCartId(cartId);
+    }
+
+    public boolean existsCartMemberByMemberIdAndCartId(Long currentMemberId, Long cartId) {
+        return cartMemberRepository.existsCartMemberByMemberIdAndCartId(currentMemberId, cartId);
+    }
+
+    public Optional<CartMember> findByMemberIdAndCartId(Long currentMemberId, Long cartId) {
+        return cartMemberRepository.findByMemberIdAndCartId(currentMemberId, cartId);
+    }
+
+    public boolean existsByMemberIdAndCartId(Long memberId, Long cartId) {
+        return cartMemberRepository.existsByMemberIdAndCartId(memberId, cartId);
     }
 }
